Rename `base` to `framework`
