--- docs/apidocs/base.rst
@@ -1,6 +1,0 @@
-.. _qiskit-experiments-base:
-
-.. automodule:: qiskit_experiments.base
-   :no-members:
-   :no-inherited-members:
-   :no-special-members:
