+++ docs/apidocs/framework.rst
@@ -1,0 +1,6 @@
+.. _qiskit-experiments-framework:
+
+.. automodule:: qiskit_experiments.framework
+   :no-members:
+   :no-inherited-members:
+   :no-special-members:
